`ursula` script uses ssh_config file, if one is found.
For cases where not all nodes are publicly addressable,
ursula now uses the ssh_config file, if one is found in
the same directory as the hosts inventory file.

This allows greater configurability of how ansible uses ssh,
such as using ProxyCommand to operate on hosts which are not
addressable from the machine running ansible.