Prevent red error output from percona playbook:
Do some things, such as checking for whether mysql is running,
in a slightly different way, to avoid red error output, as this
can mask true errors when it is a regular part of the run.

Also, break percona into a seperate playbook so that site.yml
continues to consist of only includes.