Don't show changed unnecessarily in percona tasks.
Clean up tasks which where showing up as 'changed' in
percona tasks, which weren't actually changed, by using
`changed_when:`.

The intention is to have the entire site.yml run with 0
changes after the subsequent install, as changes can be
an indication of something going wrong, and these alerts
are masked when the run always has yellow tasks.

This change introduces a dependency on ansible 1.3, to use
the changed_when keyword.