Fix glance_image:
The upstream glance_image ansible module was using glance-client v1,
which sucks, and gives http 409 when trying to create an image from
a url.

Flip to using glance-client v2, which works, and remove is_public,
which has gone away in the new api.

Verified that creating images from both urls and local files works as
expected.