Update horizon to latest trunk revision.
This was motivated by the desire to pick up this change:
  https://review.openstack.org/#/c/46804/

This prevents floating ips for other tenants from showing
up in horizon, which has been a problem for us and for customers.

Additionally, add the horizon revision as a role default,
so that it need not be specified in per-env config.