Remove conflicting lineinfiles for horizon requirements.
Ansible's lineinfile wasn't behaving as expected, and
the lineinfile for lockfile package was overwriting the one
for python-memcached.

Turns out that the lockfile requirement no longer needs to
be added, since it is already included in the upstream
horizon version we are now using (havana).

lockfile was not present in horizon's requirements on HEAD,
which was why it was added initially.