Only use memcached role on controller
We use variables from the memcached role when other roles are executing.
However, this installs memcached on servers it should not be installed on.
Using vars_files to bring in the dependant vars.  This is not ideal, b/c
it does not allow overriding, but at this point, I am starting to question
the usefulness of roles' defaults.

Ansible vars are really really annoying -- at least for me.