Decrease token expiration
This sets the token's memcached ttl.  Memcached will not return
expired keys so it is a temporary work-a-round.  A bug has already
been reported upstream [1].

The ideal configuration is not to use memcached or sql for token
backends at all.  Keystone devs are working on fully ephemeral tokens
in Icehouse.

[1] https://bugs.launchpad.net/keystone/+bug/1251123