Adding barbican/python-barbicanclient support
This change creates a new barbican role. As Barbican is still in
Stackforge, this change configures it to run from within its own
virtualenv. The service runs under uwsgi which is also part of the
specified virtualenv. Haproxy fronts this service as it would with
any other OpenStack service that we have.  The port has been moved
from 9311 to 9322 accordingly.

In addition to configuring Barbican, this change also installs
python-barbicanclient. We specifically pull directly from github due
to the fact that we have recently pushed changes upstream. These are
not yet in any of the Pypi repositories yet. Regardless, this is how
we manage the other python-*clients, so it will likely stay this way.
This client is installed into the system's python path - no
virtualenv wrangling required.