Add new crypttab module
This change adds a new crypttab module. The purpose of this module is
to support the modification of /etc/crypttab as well as provide a
mechanism to start and stop the mappings found within.

The available states are:
present - entry is found in /etc/crypttab
absent - entry is not found in /etc/crypttab
mapped - entry is found in /etc/crypttab and is actively mapped (except
  in the case where it would be impossible for us to do so - ie. no
  keyfile parameter)
unmapped - entry is removed from /etc/crypttab and mapping is unmapped
  if currently mapped.