Provide mechanism for patching nova
This change provides a mechanism for patching nova on the host. Patch
files are copied onto the host and dropped in /opt/stack/patches/nova.
Once there they are applied in lexical order.

This is intended to be a stopgap measure until we have artifacts that
we can lay down instead.