Add a default variable plugin
As it stands right now, there are a number of scenarios where we would
like to be more DRY. We have multiple environments that we support and
copying and pasting variables from one environment to the next is a
recipe for disaster. Therefore provide an optional group_vars wrapper
plugin.