Update the neutronclient revision
We are currently using an older version of the neutronclient. In that
version there is no explicit version for the cliff pip module. However,
cliff has since moved on and is no longer compatible with
neutronclient. This has been fixed upstream and as the clients do not
have version tags, pull the latest master sha.