Increase neutron agent_down_time timeout.
In one production environment, neutron was intermittently considering
dhcp-agents to be unavailable, leading to instance dhcp failures.

In this environment, dhcp-agents were checking every ~ 5-15 seconds.
The default heartbeat interval is 4 seconds, and the default agent_down_time
is 9 seconds.

It is so far known why heartbeats were not coming in as frequently as expected
in this environment, so bump agent_down_time to 20 seconds, so that
services are not incorrectly considered unavailable.