Prevent false positives in haproxy check.
Previously, the haproxy process check was looking for the string 'haproxy'
in the process command line.

This results in false positives with the introduction of LBaaS, which contains
'haproxy' in the path to its config file.

Prevent this by looking for a process which contains "haproxy.*-f $haproxy_config_file".