Fix rabbit-queues check
This check had 3 issues. The first is that it was checking again :min
and not :expected, hence would always fail.  Similarly, there is no
fall-through case when > 0 and not within the critical threshold.

Also adding a -i argument so that we can quiet down some noisy queues -
in particular, q-plugin.