Add support for the Nova Docker driver
This changeset enables support for using Nova's Docker driver. This
change is intended to use the default driver as found in the Nova
source code. This will likely get augmented to also make use of the
nova-docker driver as found on Stackforge.

These changes are do a few things:
- install the docker-registry as a controller service
- adds an ad-hoc playbook to upgrade precise for Docker
- installs docker
- adds nova user to the docker group
- updates glance to support 'docker' container format

This does not yet handle HA-ness, but that is coming. Some of the
haproxy code needs to be refactored to support this better.

newer novadocker does not need registry
seperate out docker and novadocker roles
documentation on using novadocker in docs

update for juno

updated for juno support

cruft cleanup

address @j2sol concerns

only use glance.container formats if it exists

Add support for the Nova Docker driver

This changeset enables support for using Nova's Docker driver. This
change is intended to use the default driver as found in the Nova
source code. This will likely get augmented to also make use of the
nova-docker driver as found on Stackforge.

These changes are do a few things:
- install the docker-registry as a controller service
- adds an ad-hoc playbook to upgrade precise for Docker
- installs docker
- adds nova user to the docker group
- updates glance to support 'docker' container format

This does not yet handle HA-ness, but that is coming. Some of the
haproxy code needs to be refactored to support this better.

newer novadocker does not need registry
seperate out docker and novadocker roles
documentation on using novadocker in docs