A better way to handle haproxy services template
Prevsiously we had interspersed Jinja code and variables to get a
desired haproxy config. It is much cleaner to set a baseline variable up
front, and then for the conditional services, add them as they are
enabled or disabled.

This fix also addresses some of the whitespace weirdness with the
templatized config.