Add Keystone public and admin endpoints
Some of the newer clients have begun to do service version discovery.
With that, they interrogate Keystone for the version that is supported,
and then use that data to pass all further authentication requests on to
the href provided by the query. In our case we were always returning
localhost as the next hop. This is clearly not correct.