test: Ring bell when done, add cleanup task
Ring the terminal bell when the stage of the test/run or test/setup
scripts are complete. This makes it easier to know when you are testing
and the process has completed in the background.

Split the cleanup task into its own task file and execution from within
the integration task file. This allows to easily inspect the state of
the stack when the integration test process fails and easily repeat.