haproxy: Use PPA only, workaround service script
The PPA vbernat/haproxy-1.5 now provides packages for both precise and
trusty.
However, their latest release (1.5.5) included a bad service script
which fails to stop the service. This is due to start-stop-daemon taking
a pathto a pidfile and not an actual pid via command line options. The
packaged service script is incorrect, so we use our own for now.