Integrate monitiring into associated roles
Monitoring was a stand alone playbook, that was almost like a role, but
not quite. This change moves all the monitoring tasks directly into the
associated roles for those tasks.

All the tasks have been tagged with "monitoring" and "common" which can
allow for doing a monitoring only run.

All roles that have monitoring need to depend on the monitoring-common
role, to pick up the common handler. Since there is no true common role
that everything depends on this will have to do.

nimble was put into the cinder control role, but that can easily be
moved elsewhere if desired.