Add an upgrade playbook for upgrading to juno
Upgrade from havana to juno, following
http://docs.openstack.org/openstack-ops/content/upgrades_havana-icehouse-ubuntu.html
mostly.

Our databases are already UTF-8 so that step is skipped.

Roles are applied to get new software in place, without a restart of the
services, and then services are managed in order to do database
migrations.  Some attempt is made to minimize the complete downtime of
any one service.

A new stop-services role is introduced in order to be able to interleave
it with the other roles.