Upgrade percona cluster to 5.6
This is required for juno neutron migrations.

Upgrade strategy is to do full outage instead of rolling. In a 2 db
setup, as soon as one goes down the whole cluster is considered offline
anyway. This is easier than trying to code for a 2 db node setup as well
as a 3 or more node setup.

Upgrade steps from
http://www.percona.com/doc/percona-xtradb-cluster/5.6/upgrading_guide_55_56.html#upgrading-guide