upload our test image as public
A public image allows non-admin tenants to make use of the image for
testing, like with rally. The upstream version of glance_image for a
while has supported making the image public. We don't use the locally
added cacert option any more so we don't need to carry our own
glance_image.