Upgrade percona arbiter
There is a new package for garbd to go with the rest of the percona
upgrade. This updates the arbiter role for the new package, which
includes using the shipped init script (which runs things as nobody) and
moving the config from /etc/default/garb to /etc/default/garbd.

upstart doesn't like the upstream garbd init script, so provide a
pattern for ansible to determine if the service is running or not.

The upgrade process uses compat mode on the new cluster members until
garbd is fully updated, then compat mode is removed. This should provide
for a fully rolling upgrade of the databases.

Also updates the vagrant defaults for percona versions.