Upgrade qemu/libvirt in juno
Newer versions allow online snapshots and other fun things, so lets pick
them up as we're upgrading everything else.

The versions that come from the cloud archive are built in a way that
makes them compatible with the old precise versions, particularly when
dealing with a mixed environment (for a short period of time), so we'll
use those versions.