Set gather_facts on the plays that need it
Previously, I had set gather_facts on the common play only.  This
will work if one is running a full ansible run, but if you intend to
use tags, which target specific plays, then you need to set
gather_facts: force on each play that needs it.

The measure for which plays need it are those that make use of
hostvars[$host] type logic.  This commit will address those.