Utilize running conductor service
This makes use of the nova conductor, shielding some services from
directly reading/writing to the database, instead passing messages in
rabbit to the conductor service to interact with the database. It also
provides translations for RPC message objects when an object is recieved
by a nova service and that service cannot understand the version of the
object.

Conductor service will enable better orchestrated upgrades of nova to
reduce downtime as much as possible and allows for mixed versions of
services running, and allows for database to be migrated before
upgrading nova-compute.