Generated new ssl keys with SaN
SubjectAltName is required in never versions of ssl clients. Without SaN
the tools will complain, a lot. So this is a newly generated self signed
cert with a proper SaN. Subject details were pulled from the previous
cert.

Also take this opportunity to upgrade to a sha256 cert instead of a
sha1.  http://blog.chromium.org/2014/09/gradually-sunsetting-sha-1.html