Only bootstrap keystone once
Since we will be generating a token and restarting services in order to
do the boot strap, we only want to boot strap once, instead of every
time we run ursula on an environment. This commit allows for that by
using a "keystone_configured" variable that is established at runtime by
looking for a touched file in /etc/keystone.