Upgrade ironic services
Ironic also runs nova-compute on the controller, so some manipulation of
it is necessary, and we don't want to run nova-data on compute when
ironic is at play.

This also required tweaking the ironic role to allow for an upgrade
process where services are stopped before the db is synced.