Update cert generation to sha256 and include SAN
SSL clients are starting to throw warnings on missing subjectAltName,
and on sha1 certificates. So update our routine to add a SAN and to
generate sha256 certs. I couldn't find a way to express our SAN via x509
outside of a config file, so now we generate a config file too for usage
during the generation.

This also updates the cert/key listed in the example environment to be
invalid, since it gets generated at runtime, and we shouldn't have valid
key/cert sitting anywhere.