WIP branch release validation playbook
This playbook tests a number of connectivity things with nova and
neutron and instances. It tests instances being on different physical
nodes and being able to ping each other, and tests instances being able
to ping the internet with a floating IP assigned and without.

Some modifications were necessary to nova_group and nova_compute local
modules which should be able to go away if/when we pick up ansible 1.8.

Playbook prompts the user to do some manual action as some of the
testing isn't quite possible in an automated fashion (eg running
commands on an instance that no longer has a floating IP assigned).