Allow tuning haproxy bufsize
In larger capacity environments it's possible to have nova generate
requests to neutron that are larger than the haproxy allowed default of
1/2 bufsize (since it has to rewrite). Tuning this up will consume more
memory and reduce capacity though, so it should only be done on larger
systems. This provides an easy way to bump it up from the default when
deemed necessary.