Support openstack cinder volume encryption
This makes use of the fixed key attribute to provide nova a hex key to
use with LUKS to encrypt/decrypt cinder volumes. Cinder doesn't need the
key, it just has to /have/ a key configured so that it'll allow
encryption flags on volumes.

Because the key is written to the config, nova has to be separated from
cinder-volume, so if a fixed_key is defined do not do the small loopback
volume device.

Encryption can be turned on simply by defining a fixed_key and an
appropriate encrypted_volume type in an environment config.