Add prereboot tag and workaround grub-common crashkernel config
Tag for items which should be updated before a reboot. Intention is that
this can be run from -master or another recent branch on an existing
stack before scheduled reboots occur on that stack.

In Precise grub-common includes an automatic crashkernel config, but it
isn't large enough for our updater kernels and configurations. Rather
than removing grub-common or backporting from a later release, just fix
the size of the crashkernel reservation to 256M.

Conflicts:
	roles/common/tasks/kernel-tuning.yml