Add support for vagrant libvirt provider
Turns out, virtualbox doesn't work on my machine because I alredy
run KVM instances for other reasons. This adds support for using
the libvirt provider to fix that.

How to use:
Install the vagrant libvirt provider plugin
  (see https://github.com/pradels/vagrant-libvirt - install deps first!)
Install the vargant mutate plugin:
  vagrant plugin install vagrant-mutate
Download the existing virtualbox image
Use vagrant mutate on the existing box image:
  vagrant mutate ursula-precise libvirt
Run as normal!

Really, that's it. The rest of the changes are covered by this
commit. Everything _seems_ to work for now, and it makes life
better for non-Mac people :D