Explicitly remove package symlink on upgrade
The package role will not update the symlink if it already exists. For
upgrades that's no good, we want a new symlink. This removal will force
the symlink to be re-created.

Also use the new suggested task format since it's required for some of
our tasks. max_fail_percentage is set to 0 because that was more
appropriate than 1.