Update swift keystoneauth options
In our previous releases, we were not using keystoneauth to validate
user role, so anybody with a valid token was allowed to use Swift. Now
we'll allow _member_ in, to match that functionality.

Also update the authtoken config to remove the old format of urls.

Temporarly skip ssl validation to keystone.