Remove definition of instance_prefix
This is being loaded as a vars_file, which has a higher precidence than
inventory variables. Thus the inventory variable that was setting the
actual prefix we wanted was getting overridden by this setting here to
just "test".

Also move the inventory to use the localhost entry, as when it's part of
the inventory it picks up the variable data we're adding. Otherwise the
variable data was lost.