Adding provision tooling to Vagrantfile to install items in bootstrap.sh rather than relying on ansible for it.
Removed the ansible bits at the bottom of the Vagrantfile that provisioned out common workstation tools (wget, vim etc) and replaced that call
with a vagrant provisioning call to bootstrap.sh. Also added bootstrap.sh. This was done because the machine that was commonly used to do this
work and run the vagrant workstation commonly did not have ansible installed and would error out at this point.