Make keystone tech spec compliant
Change the file permissions of the /etc/keystone/ files to 640, the owner
to keystone, and the group to keystone. Note that policy.json must remain
644 so that Horizon can still check for sidebar policies. This change is in
roles/keystone/tasks/main.yml.

Change the file permissions of the logs in /var/log/keystone/ to 644 so
log stash can still function and change the owner and group to keystone.
This change is in roles/keystone/tasks/main.yml.

Change the default_log_level to lower the keystonemiddleware logging level.
This change is in roles/keystone/templates/etc/keystone/keystone.conf.