Install python-prettytable for client usage
If the package is not installed, pip will pull in python-prettytable
from pypi. Installation this way can lead to prettytable packages that
are only readable by root, breaking non-root usage of clients or other
things like that. Work around this by installing the package from apt to
satisfy the dependency.