library: sensu modules use default, metrics handlers
Since we are moving to Flapjack.io as part of the
alerting pipeline, we should use a default and metrics
handlers instead of explicitly using pagerduty and graphite.

This allows us to easily control which handlers are actually
used as part of a set of handlers on the sensu-server.
Currently, default set is pagerduty and flapjack_http, and the
metrics handler is graphite.

This gives us flexibility rather than updating ursula proper with
our internal requirements and systems.