sensu modules action plugins
Action plugins allow for variables or facts to be read and
provided to a module during the normal task execution.

Specifically, this allows for a default value to be set like
a normal ansible variable is set within roles, defaults, group_vars,
or host_vars. This follows the normal variable and fact merging.

A specific override of a single task is still supported.