Move keystone defaults to a data only role
We need the keystone defaults in multiple roles, but we don't want to go
through all the keystone tasks in each role, so move the defaults to its
own role.

This should fix the keystone-setup task when it writes out
keystone-paste.ini so that it has all the right information for the
conditionals.

Also turn back on our protection against attempting to reconfigure
keystone in subsequent runs.