Generalize references to files in other roles
Errors occur while attempting to use ursula with a playbook
similiar to site.yml that does not exist in the root ursula
directory.  The errors are do to references to files in one
role that exist in another role. The file reference assumes
the roles directory is in the same directory as the playbook.

The change made uses the role_path variable to find files that
are referenced by other roles.