Speed up ansible by tweaking ssh
This introduces our own connection plugin again.

First, we are disabling host key checking at the Ansible level, but we
don't want to disable it from the raw ssh level, so we tweak the
connection plugin to not take an opinion on that. Environments can
express an ssh_config file and decide whether or not to do host key
checking there.

Second, we don't let Ansible read through known_hosts, we're offloading
that work to ssh itself.

Finally, we don't want to attempt Kerberos logins.