Add a callback to fail on unreachable hosts
In 1.9, particularly in gather_facts, an unreachable host doesn't count
as a failed host for any_errors_fatal type settings. So this callback
will trigger when a host is unreachable and just truncate the play hosts
to an empty set, which will abort the play on the very next task.