Serialize cinder service (re)starts
When using a common host name for shared backend scenarios (like ceph),
starting two services at once with the same name can lead to a race
condition. We want to serialize at least the initial start up of the
service, but since that could be via a handler OR the role task we need
to serialize both.