Update upgrade playbook for kilo -> mitaka
Suprisingly little had to chance. Removal of a bunch of no longer needed
upgrade bits, and re-ordering of services, and adding forced fact
gathering in a few places.

Also a failed_when conditional in Heat had to be updated as the old
keystone output didn't match the new keystone output.