Introduce oslo middleware for http proxy (nova)
This is new oslo middleware to automatically translate from an http(s)
proxy to the wsgi application so that all the right references get sent
back. This replaces the one-off handling of the HTTPS proxy header.

This is just for nova to prove the model, before applying to all
services.

https://review.openstack.org/#/c/227868/4