Generate our own host uuid for libvirt
This does two things:
1) Ensure a uuid exists in /etc/machine-id
2) Write value from machine-id into libvirt config as host_uuid

host_uuid should always be unique, as it's used to validate a live
migration target.

Nova exposes to the guest some information about the host, including the
uuid. The nova default is to read /etc/machine-id and fallback to what
libvirt shows. Since we need to store a generated uuid somewhere, we're
picking /etc/machine-id for this storage.