rework restarting services on code changes
The old method didn't work if the same hosts got ran through
package/source installs more than once. The first time through would
result in a task with a changed status, but a second time through it
would not, so some services were not getting killed or restarted. This
introduces a single variable that is set to true when change happens,
and left alone when change does not.