Remove default availability zone settings
Turns out that if nothing is specified, the scheduler will just toss it
at any available compute and not try to filter by AZ. This is most
likely what we want for our stacks.

However we do want to retain the ability for for a site to override the
default, which can be done in site config.