Use upstream ansible modules for keystone setup
We get access to os_user, os_keystone_role, and os_user_role now so lets
use them. Also we need to set up the admin user to have admin role on
the default domain, which is a new keystonev3 thing.

Given these modules come in Ansible 2.1.0.0 we need to require that
version.

Also bump our version of shade to handle those modules.