Unipin (most) clients
We no longer rely on the command line in our automation, and we don't
rely on the libraries as much any more given use of shade. Unpinning
these seems to make the clients work again, including openstackclient,
which wasn't previously working.

Keystone client is pinned under 3.0.0 as the newer one does break a few
things.