Work around shade domain bugs in heat setup
Shade backed modules don't support domains very well, when Keystone has
domain specific configuration enabled, which prevents keystone from
returning all users when doing an API listing.

To work around this, we track when we add the heat admin user, and only
if we add that user do we attempt to use the command line to assign a
role. The os_user_role module cannot work here as it can't find the user
we're trying to manipulate, because it doesn't look in the right domain.

This work around can go away when things like
https://storyboard.openstack.org/?#!/story/2000615 are fixed.