data-driven-checks: cross-host keepalive alerts
When an openstack host goes down, especially if it is a controller, it
often triggers an influx of alerts. This can cause confusion and hide
the underyling cause of the problem to the operator.

To rememdy this problem, sensu checks (and our infrastructure) supports
the notion of "check dependencies" which can be used to automatically
"silence" an alert. The "silenced" alerts are still seen in Uchiwa, but
are NOT forwarded to our paging system.

Along with this change, we are moving to "data driven checks" which
allow for more fields to be added to checks, as well as fields to be
easily overriden.