Support for pip upper constraints
Each project can specify an upper constraints file to fetch and use to
constrain the deps and project itself.

A project can still opt out of being constrained.

This also has the byproduct of fixing source installs with a pypi mirror
(which was broken), and updating pip inside the virtualenv (which was
really really old).

Change-Id: Idb02f7bb3071f3cab336bf473d1c70ca1a7ede46