Fix identity panel for project_admin (#1982)
We are enabling a custom keystone policy file for horizon because
the horizon keystone policy and keystone policy need to be slightly
different when supporting the project admin. The horizon keystone policy
file requires project admin to be part of the admin required rule.
The project_id was also being used in the policy where project scope
did not matter. This has been corrected.
This fixes a bug where horizon fails to load parts of horizon
if the user has the project_admin role.