Work around bugs in os_user_facts (#2100)
The contents of the auth dict get marked as no_log, which apparently
transforms the value of the data returned by os_user_facts. By moving
the 'admin' words out of the auth dict, the correct values come through.

In addition, the attribute where the name is found is "name" not
"username". Finally, it's a fact module so the facts are integrated
directly into the host facts, so we can access them without a register.

Change-Id: I9e28aa566c76372c31e16d4c0c138dc77fc2b146