Add support for SAML federation for Keystone (#2151)
This adds support for SAML federation. Most of the work for
SAML federation was there because of the Keystone to Keystone Federation
feature. This commit makes the shibboleth configuration more
flexible to support either K2K federation or an external SAML IDP.

We will currently support only one Identity provider for SAML. Either
a K2K IDP or another external SAML IDP. We plan on supporting multiple
SAML IDPs in the future so we have changed the data structure for
SAML SPs into a list.