Add tasks to upgrade percona to latest minor
This is needed to upgrade our percona to the latest minor release
offered. Installing the package will cause the DB to restart, so we want
to take care that the cluster is in a healthy state before we move on.
Also, becuase this is risky, we want to add an additional run-time flag
of 'dbupgrade' to determine whether or not these tasks run.

Change-Id: I0dbdf21a828325a2dbac08c2c812f97b7f779ec2