make admin keyring owned by ceph on cpm (#2251)
In 3.0.x(ceph version is Hammer), admin keyring is owned by root on cpm. 
While in 3.1.0(ceph version is Jewel), admin keyring is owned by ceph
user.
Admin keyring is generated by ceph program, but it's used by users.
Ceph program doesn't use it. So it doesn't matter who is the owner of
admin keyring.
But it's better to make the owner consistent.